Strip/relocate old code
