+++ Makefile
@@ -10,14 +10,14 @@ clean:
 
 check: ci test
 
-test_engine:
-	colcon test --event-handlers console_direct+ --packages-select sgengine sgengine_messages
+test:
+	colcon test --event-handlers console_direct+ --packages-skip python-steamcontroller
 
 ci:
 	python3 -m black ./sgengine/sgengine
-	python3 -m pylint ./sgengine/sgengine --ignore-paths=sgengine/sgengine/hardware/utils --ignore-paths=sgengine/sgengine/aruco --disable=C0303,C0114,W0401,R0902,R0904,C0305,W0703,C0200,R0913,R0914,W0107,W1203,W0246,W0511,R0801
-	python3 -m flake8 --exclude=sgengine/sgengine/hardware/utils --exclude=sgengine/sgengine/aruco ./sgengine/sgengine --count --select=E9,F63,F7,F82,F401,W292,E275,F403 --extend-ignore=W293,W291,E303,E203,W503 --show-source --statistics
-	python3 -m flake8 --exclude=sgengine/sgengine/hardware/utils --exclude=sgengine/sgengine/aruco ./sgengine/sgengine --count --exit-zero --max-complexity=10 --max-line-length=140 --extend-ignore=E203 --statistics
+	python3 -m pylint ./sgengine/sgengine --disable=C0303,C0114,W0401,R0902,R0904,C0305,W0703,C0200,R0913,R0914,W0107,W1203,W0246,W0511,R0801
+	python3 -m flake8 ./sgengine/sgengine --count --select=E9,F63,F7,F82,F401,W292,E275,F403 --extend-ignore=W293,W291,E303,E203,W503 --show-source --statistics
+	python3 -m flake8 ./sgengine/sgengine --count --exit-zero --max-complexity=10 --max-line-length=140 --extend-ignore=E203 --statistics
 	@echo "DONE - CI PASSED"
 
 help:
