+++ launch/auton_control.launch.py
@@ -5,8 +5,13 @@ def generate_launch_description():
     return LaunchDescription([
         Node(
             package='sgengine',
+            executable='pico',
+            name='pico'
+        ),
+        Node(
+            package='sgengine',
             executable='odometry',
             name='odometry'
-        ),
+        )
     ])
