+++ sgengine/package.xml
@@ -8,13 +8,7 @@
   <license>MIT</license>
 
   <buildtool_depend>ament_python</buildtool_depend>
-
-  <test_depend>ament_copyright</test_depend>
-  <test_depend>ament_flake8</test_depend>
-  <test_depend>ament_pep257</test_depend>
-  <test_depend>python3-pytest</test_depend>
-  <test_depend>ament_mypy</test_depend>
-
+  
   <depend>rclpy</depend>
   <depend>std_msgs</depend>
   <depend>depthai</depend>
