--- sgengine/sgengine/configuration/__init__.py
@@ -1,3 +1,0 @@
-# from .config import Config
-
-# __all__ = ["Config"]
