--- sgengine/sgengine/gui/__init__.py
@@ -1,3 +1,0 @@
-from .gui_node import GUINode
-
-__all__ = ["GUINode"]
