--- sgengine/sgengine/hardware/__init__.py
@@ -1,4 +1,0 @@
-from .hardware_node import HardwareNode
-from .oakds2 import OakDS2Node, OakDS2
-
-__all__ = ["HardwareNode", "OakDS2", "OakDS2Node"]
