--- sgengine/sgengine/hardware/oakds2/__init__.py
@@ -1,4 +1,0 @@
-from .oakds2 import OakDS2
-from .oakds2_node import OakDS2Node
-
-__all__ = ["OakDS2", "OakDS2Node"]
