--- sgengine/sgengine/map/map_node.py
@@ -1,15 +1,0 @@
-from ..abstract_node import AbstractNode
-
-
-class MapNode(AbstractNode):
-    """Node for handling map functionality"""
-
-    def __init__(self, name: str, *args, **kwargs) -> None:
-        super().__init__(name, *args, **kwargs)
-
-        # TODO: define map variables
-        pass
-
-    def _main(self):
-        # TODO: define pubs and sub
-        pass
