--- sgengine/sgengine/pathfinding/pathfinding_node.py
@@ -1,15 +1,0 @@
-from ..abstract_node import AbstractNode
-
-
-class PathfindingNode(AbstractNode):
-    """Node for handling pathfinding"""
-
-    def __init__(self, name: str, *args, **kwargs) -> None:
-        super().__init__(name, *args, **kwargs)
-
-        # TODO: define attributes needed for pathfinding
-        pass
-
-    def _main(self):
-        # TODO: define pubs and subs for pathfinding
-        pass
